+++ src/main/java/net/earthcomputer/clientcommands/command/commands/CHelpCommand.java
@@ -1,4 +1,4 @@
-package net.earthcomputer.clientcommands.command;
+package net.earthcomputer.clientcommands.command.commands;
 
 import com.google.common.collect.Iterables;
 import com.mojang.brigadier.CommandDispatcher;
