+++ src/main/java/net/earthcomputer/clientcommands/command/commands/MoteCommand.java
@@ -1,4 +1,4 @@
-package net.earthcomputer.clientcommands.command;
+package net.earthcomputer.clientcommands.command.commands;
 
 import com.mojang.brigadier.CommandDispatcher;
 import net.minecraft.client.MinecraftClient;
